requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.1
django-filter==2.3.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
idna==2.10
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
pylint==2.6.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
